@@ -1,6 +1,6 @@
 [package]
 name = "just"
-version = "1.24.0"
+version = "1.26.0"
 authors = ["Casey Rodarmor <casey@rodarmor.com>"]
 autotests = false
 categories = ["command-line-utilities", "development-tools"]
@@ -12,33 +12,37 @@ keywords = ["command-line", "task", "runner", "development", "utility"]
 license = "CC0-1.0"
 readme = "crates-io-readme.md"
 repository = "https://github.com/casey/just"
-rust-version = "1.63"
+rust-version = "1.70"
 
 [workspace]
-members = [".", "bin/ref-type", "bin/generate-book", "bin/update-contributors"]
+members = [".", "crates/*"]
 
 [dependencies]
 ansi_term = "0.12.0"
-atty = "0.2.0"
 blake3 = { version = "1.5.0", features = ["rayon", "mmap"] }
 camino = "1.0.4"
-clap = { version = "2.33.0", features = ["wrap_help"] }
+clap = { version = "4.0.0", features = ["env", "wrap_help"] }
+clap_complete = "4.0.0"
+clap_mangen = "0.2.20"
 ctrlc = { version = "3.1.1", features = ["termination"] }
 derivative = "2.0.0"
 dirs = "5.0.1"
 dotenvy = "0.15"
 edit-distance = "2.0.0"
 env_logger = "0.11.0"
-heck = "0.4.0"
+heck = "0.5.0"
 lexiclean = "0.0.1"
 libc = "0.2.0"
 log = "0.4.4"
 num_cpus = "1.15.0"
-regex = "1.5.4"
+percent-encoding = "2.3.1"
+rand = "0.8.5"
+regex = "1.10.4"
 semver = "1.0.20"
 serde = { version = "1.0.130", features = ["derive", "rc"] }
 serde_json = "1.0.68"
 sha2 = "0.10"
+shellexpand = "3.1.0"
 similar = { version = "2.1.0", features = ["unicode"] }
 snafu = "0.8.0"
 strum = { version = "0.26.0", features = ["derive"] }
@@ -54,7 +58,6 @@ executable-path = "1.0.0"
 pretty_assertions = "1.0.0"
 temptree = "0.2.0"
 which = "6.0.0"
-yaml-rust = "0.4.5"
 
 [lib]
 doctest = false
@@ -64,13 +67,6 @@ path = "src/main.rs"
 name = "just"
 test = false
 
-[features]
-# No features are active by default.
-default = []
-# The `help4help2man` feature modifies the message produced by `--help`
-# so that `help2man` produces a reasonable man page.
-help4help2man = []
-
 # The public documentation is minimal and doesn't change between
 # platforms, so we only build them for linux on docs.rs to save
 # their build machines some cycles.
@@ -79,6 +75,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [profile.release]
 lto = true
+codegen-units = 1
 
 [[test]]
 name = "integration"
