@@ -38,16 +38,16 @@ pub(crate) struct Lexer<'src> {
 impl<'src> Lexer<'src> {
   /// Lex `src`
   pub(crate) fn lex(path: &'src Path, src: &'src str) -> CompileResult<'src, Vec<Token<'src>>> {
-    Lexer::new(path, src).tokenize()
+    Self::new(path, src).tokenize()
   }
 
   #[cfg(test)]
   pub(crate) fn test_lex(src: &'src str) -> CompileResult<'src, Vec<Token<'src>>> {
-    Lexer::new("justfile".as_ref(), src).tokenize()
+    Self::new("justfile".as_ref(), src).tokenize()
   }
 
   /// Create a new Lexer to lex `src`
-  fn new(path: &'src Path, src: &'src str) -> Lexer<'src> {
+  fn new(path: &'src Path, src: &'src str) -> Self {
     let mut chars = src.chars();
     let next = chars.next();
 
@@ -57,7 +57,7 @@ impl<'src> Lexer<'src> {
       line: 0,
     };
 
-    Lexer {
+    Self {
       indentation: vec![""],
       tokens: Vec::new(),
       token_start: start,
@@ -231,11 +231,8 @@ impl<'src> Lexer<'src> {
     // The width of the error site to highlight depends on the kind of error:
     let length = match kind {
       UnterminatedString | UnterminatedBacktick => {
-        let kind = match StringKind::from_token_start(self.lexeme()) {
-          Some(kind) => kind,
-          None => {
-            return self.internal_error("Lexer::error: expected string or backtick token start")
-          }
+        let Some(kind) = StringKind::from_token_start(self.lexeme()) else {
+          return self.internal_error("Lexer::error: expected string or backtick token start");
         };
         kind.delimiter().len()
       }
@@ -282,11 +279,7 @@ impl<'src> Lexer<'src> {
 
   /// True if `c` can be a continuation character of an identifier
   fn is_identifier_continue(c: char) -> bool {
-    if Self::is_identifier_start(c) {
-      return true;
-    }
-
-    matches!(c, '0'..='9' | '-')
+    Self::is_identifier_start(c) || matches!(c, '0'..='9' | '-')
   }
 
   /// Consume the text and produce a series of tokens
@@ -817,9 +810,7 @@ impl<'src> Lexer<'src> {
   /// Cooked string: "[^"]*" # also processes escape sequences
   /// Raw string:    '[^']*'
   fn lex_string(&mut self) -> CompileResult<'src> {
-    let kind = if let Some(kind) = StringKind::from_token_start(self.rest()) {
-      kind
-    } else {
+    let Some(kind) = StringKind::from_token_start(self.rest()) else {
       self.advance()?;
       return Err(self.internal_error("Lexer::lex_string: invalid string start"));
     };
@@ -1028,7 +1019,7 @@ mod tests {
             length,
             path: "justfile".as_ref(),
           },
-          kind: Box::new(kind),
+          kind: kind.into(),
         };
         assert_eq!(have, want);
       }
