@@ -13,9 +13,16 @@ pub(crate) enum Error<'src> {
     min: usize,
     max: usize,
   },
+  Assert {
+    message: String,
+  },
   Backtick {
     token: Token<'src>,
     output_error: OutputError,
+  },
+  CacheDirIo {
+    io_error: io::Error,
+    path: PathBuf,
   },
   ChooserInvoke {
     shell_binary: String,
@@ -139,8 +146,14 @@ pub(crate) enum Error<'src> {
     line_number: Option<usize>,
     signal: i32,
   },
-  TmpdirIo {
+  StdoutIo {
+    io_error: io::Error,
+  },
+  TempdirIo {
     recipe: &'src str,
+    io_error: io::Error,
+  },
+  TempfileIo {
     io_error: io::Error,
   },
   Unknown {
@@ -256,6 +269,9 @@ impl<'src> ColorDisplay for Error<'src> {
           write!(f, "Recipe `{recipe}` got {found} {count} but takes at most {max}")?;
         }
       }
+      Assert { message }=> {
+        write!(f, "Assert failed: {message}")?;
+      }
       Backtick { output_error, .. } => match output_error {
         OutputError::Code(code) => write!(f, "Backtick failed with exit code {code}")?,
         OutputError::Signal(signal) => write!(f, "Backtick was terminated by signal {signal}")?,
@@ -266,6 +282,9 @@ impl<'src> ColorDisplay for Error<'src> {
             _ => write!(f, "Backtick could not be run because of an IO error while launching the shell:\n{io_error}"),
           }?,
         OutputError::Utf8(utf8_error) => write!(f, "Backtick succeeded but stdout was not utf8: {utf8_error}")?,
+      }
+      CacheDirIo { io_error, path } => {
+        write!(f, "I/O error in cache dir `{}`: {io_error}", path.display())?;
       }
       ChooserInvoke { shell_binary, shell_arguments, chooser, io_error} => {
         let chooser = chooser.to_string_lossy();
@@ -370,8 +389,7 @@ impl<'src> ColorDisplay for Error<'src> {
         }?;
       }
       Load { io_error, path } => {
-        let path = path.display();
-        write!(f, "Failed to read justfile at `{path}`: {io_error}")?;
+        write!(f, "Failed to read justfile at `{}`: {io_error}", path.display())?;
       }
       MissingImportFile { .. } => write!(f, "Could not find source file for import.")?,
       MissingModuleFile { module } => write!(f, "Could not find source file for module `{module}`.")?,
@@ -397,9 +415,15 @@ impl<'src> ColorDisplay for Error<'src> {
           write!(f, "Recipe `{recipe}` was terminated by signal {signal}")?;
         }
       }
-      TmpdirIo { recipe, io_error } => {
+      StdoutIo { io_error } => {
+        write!(f, "I/O error writing to stdout: {io_error}?")?;
+      }
+      TempdirIo { recipe, io_error } => {
         write!(f, "Recipe `{recipe}` could not be run because of an IO error while trying to create a temporary \
-                   directory or write a file to that directory`:{io_error}")?;
+                   directory or write a file to that directory: {io_error}")?;
+      }
+      TempfileIo { io_error } => {
+        write!(f, "Tempfile I/O error: {io_error}")?;
       }
       Unknown { recipe, line_number} => {
         if let Some(n) = line_number {
