@@ -1,12 +1,12 @@
 use super::*;
 
 pub(crate) struct Evaluator<'src: 'run, 'run> {
-  assignments: Option<&'run Table<'src, Assignment<'src>>>,
-  config: &'run Config,
-  dotenv: &'run BTreeMap<String, String>,
-  scope: Scope<'src, 'run>,
-  settings: &'run Settings<'run>,
-  search: &'run Search,
+  pub(crate) assignments: Option<&'run Table<'src, Assignment<'src>>>,
+  pub(crate) config: &'run Config,
+  pub(crate) dotenv: &'run BTreeMap<String, String>,
+  pub(crate) scope: Scope<'src, 'run>,
+  pub(crate) settings: &'run Settings<'run>,
+  pub(crate) search: &'run Search,
 }
 
 impl<'src, 'run> Evaluator<'src, 'run> {
@@ -68,30 +68,23 @@ impl<'src, 'run> Evaluator<'src, 'run> {
       Expression::Call { thunk } => {
         use Thunk::*;
 
-        let context = FunctionContext {
-          dotenv: self.dotenv,
-          invocation_directory: &self.config.invocation_directory,
-          search: self.search,
-        };
-
         match thunk {
-          Nullary { name, function, .. } => {
-            function(&context).map_err(|message| Error::FunctionCall {
-              function: *name,
-              message,
-            })
-          }
+          Nullary { name, function, .. } => function(self).map_err(|message| Error::FunctionCall {
+            function: *name,
+            message,
+          }),
           Unary {
             name,
             function,
             arg,
             ..
-          } => function(&context, &self.evaluate_expression(arg)?).map_err(|message| {
-            Error::FunctionCall {
+          } => {
+            let arg = self.evaluate_expression(arg)?;
+            function(self, &arg).map_err(|message| Error::FunctionCall {
               function: *name,
               message,
-            }
-          }),
+            })
+          }
           UnaryOpt {
             name,
             function,
@@ -104,7 +97,23 @@ impl<'src, 'run> Evaluator<'src, 'run> {
               None => None,
             };
 
-            function(&context, &a, b.as_deref()).map_err(|message| Error::FunctionCall {
+            function(self, &a, b.as_deref()).map_err(|message| Error::FunctionCall {
+              function: *name,
+              message,
+            })
+          }
+          UnaryPlus {
+            name,
+            function,
+            args: (a, rest),
+            ..
+          } => {
+            let a = self.evaluate_expression(a)?;
+            let mut rest_evaluated = Vec::new();
+            for arg in rest {
+              rest_evaluated.push(self.evaluate_expression(arg)?);
+            }
+            function(self, &a, &rest_evaluated).map_err(|message| Error::FunctionCall {
               function: *name,
               message,
             })
@@ -114,15 +123,14 @@ impl<'src, 'run> Evaluator<'src, 'run> {
             function,
             args: [a, b],
             ..
-          } => function(
-            &context,
-            &self.evaluate_expression(a)?,
-            &self.evaluate_expression(b)?,
-          )
-          .map_err(|message| Error::FunctionCall {
-            function: *name,
-            message,
-          }),
+          } => {
+            let a = self.evaluate_expression(a)?;
+            let b = self.evaluate_expression(b)?;
+            function(self, &a, &b).map_err(|message| Error::FunctionCall {
+              function: *name,
+              message,
+            })
+          }
           BinaryPlus {
             name,
             function,
@@ -131,13 +139,11 @@ impl<'src, 'run> Evaluator<'src, 'run> {
           } => {
             let a = self.evaluate_expression(a)?;
             let b = self.evaluate_expression(b)?;
-
             let mut rest_evaluated = Vec::new();
             for arg in rest {
               rest_evaluated.push(self.evaluate_expression(arg)?);
             }
-
-            function(&context, &a, &b, &rest_evaluated).map_err(|message| Error::FunctionCall {
+            function(self, &a, &b, &rest_evaluated).map_err(|message| Error::FunctionCall {
               function: *name,
               message,
             })
@@ -147,16 +153,15 @@ impl<'src, 'run> Evaluator<'src, 'run> {
             function,
             args: [a, b, c],
             ..
-          } => function(
-            &context,
-            &self.evaluate_expression(a)?,
-            &self.evaluate_expression(b)?,
-            &self.evaluate_expression(c)?,
-          )
-          .map_err(|message| Error::FunctionCall {
-            function: *name,
-            message,
-          }),
+          } => {
+            let a = self.evaluate_expression(a)?;
+            let b = self.evaluate_expression(b)?;
+            let c = self.evaluate_expression(c)?;
+            function(self, &a, &b, &c).map_err(|message| Error::FunctionCall {
+              function: *name,
+              message,
+            })
+          }
         }
       }
       Expression::StringLiteral { string_literal } => Ok(string_literal.cooked.clone()),
@@ -171,22 +176,11 @@ impl<'src, 'run> Evaluator<'src, 'run> {
         Ok(self.evaluate_expression(lhs)? + &self.evaluate_expression(rhs)?)
       }
       Expression::Conditional {
-        lhs,
-        rhs,
+        condition,
         then,
         otherwise,
-        operator,
       } => {
-        let lhs_value = self.evaluate_expression(lhs)?;
-        let rhs_value = self.evaluate_expression(rhs)?;
-        let condition = match operator {
-          ConditionalOperator::Equality => lhs_value == rhs_value,
-          ConditionalOperator::Inequality => lhs_value != rhs_value,
-          ConditionalOperator::RegexMatch => Regex::new(&rhs_value)
-            .map_err(|source| Error::RegexCompile { source })?
-            .is_match(&lhs_value),
-        };
-        if condition {
+        if self.evaluate_condition(condition)? {
           self.evaluate_expression(then)
         } else {
           self.evaluate_expression(otherwise)
@@ -198,32 +192,53 @@ impl<'src, 'run> Evaluator<'src, 'run> {
         lhs: Some(lhs),
         rhs,
       } => Ok(self.evaluate_expression(lhs)? + "/" + &self.evaluate_expression(rhs)?),
+      Expression::Assert { condition, error } => {
+        if self.evaluate_condition(condition)? {
+          Ok(String::new())
+        } else {
+          Err(Error::Assert {
+            message: self.evaluate_expression(error)?,
+          })
+        }
+      }
     }
   }
 
+  fn evaluate_condition(&mut self, condition: &Condition<'src>) -> RunResult<'src, bool> {
+    let lhs_value = self.evaluate_expression(&condition.lhs)?;
+    let rhs_value = self.evaluate_expression(&condition.rhs)?;
+    let condition = match condition.operator {
+      ConditionalOperator::Equality => lhs_value == rhs_value,
+      ConditionalOperator::Inequality => lhs_value != rhs_value,
+      ConditionalOperator::RegexMatch => Regex::new(&rhs_value)
+        .map_err(|source| Error::RegexCompile { source })?
+        .is_match(&lhs_value),
+    };
+    Ok(condition)
+  }
+
   fn run_backtick(&self, raw: &str, token: &Token<'src>) -> RunResult<'src, String> {
+    self
+      .run_command(raw, &[])
+      .map_err(|output_error| Error::Backtick {
+        token: *token,
+        output_error,
+      })
+  }
+
+  pub(crate) fn run_command(&self, command: &str, args: &[&str]) -> Result<String, OutputError> {
     let mut cmd = self.settings.shell_command(self.config);
-
-    cmd.arg(raw);
-
+    cmd.arg(command);
+    cmd.args(args);
     cmd.current_dir(&self.search.working_directory);
-
     cmd.export(self.settings, self.dotenv, &self.scope);
-
     cmd.stdin(Stdio::inherit());
-
     cmd.stderr(if self.config.verbosity.quiet() {
       Stdio::null()
     } else {
       Stdio::inherit()
     });
-
-    InterruptHandler::guard(|| {
-      output(cmd).map_err(|output_error| Error::Backtick {
-        token: *token,
-        output_error,
-      })
-    })
+    InterruptHandler::guard(|| output(cmd))
   }
 
   pub(crate) fn evaluate_line(
