@@ -8,6 +8,11 @@ use super::*;
 /// The parser parses both values and expressions into `Expression`s.
 #[derive(PartialEq, Debug, Clone)]
 pub(crate) enum Expression<'src> {
+  /// `assert(condition, error)`
+  Assert {
+    condition: Condition<'src>,
+    error: Box<Expression<'src>>,
+  },
   /// `contents`
   Backtick {
     contents: String,
@@ -20,13 +25,11 @@ pub(crate) enum Expression<'src> {
     lhs: Box<Expression<'src>>,
     rhs: Box<Expression<'src>>,
   },
-  /// `if lhs == rhs { then } else { otherwise }`
+  /// `if condition { then } else { otherwise }`
   Conditional {
-    lhs: Box<Expression<'src>>,
-    rhs: Box<Expression<'src>>,
+    condition: Condition<'src>,
     then: Box<Expression<'src>>,
     otherwise: Box<Expression<'src>>,
-    operator: ConditionalOperator,
   },
   /// `(contents)`
   Group { contents: Box<Expression<'src>> },
@@ -50,27 +53,23 @@ impl<'src> Expression<'src> {
 impl<'src> Display for Expression<'src> {
   fn fmt(&self, f: &mut Formatter) -> Result<(), fmt::Error> {
     match self {
-      Expression::Backtick { token, .. } => write!(f, "{}", token.lexeme()),
-      Expression::Join { lhs: None, rhs } => write!(f, "/ {rhs}"),
-      Expression::Join {
+      Self::Assert { condition, error } => write!(f, "assert({condition}, {error})"),
+      Self::Backtick { token, .. } => write!(f, "{}", token.lexeme()),
+      Self::Join { lhs: None, rhs } => write!(f, "/ {rhs}"),
+      Self::Join {
         lhs: Some(lhs),
         rhs,
       } => write!(f, "{lhs} / {rhs}"),
-      Expression::Concatenation { lhs, rhs } => write!(f, "{lhs} + {rhs}"),
-      Expression::Conditional {
-        lhs,
-        rhs,
+      Self::Concatenation { lhs, rhs } => write!(f, "{lhs} + {rhs}"),
+      Self::Conditional {
+        condition,
         then,
         otherwise,
-        operator,
-      } => write!(
-        f,
-        "if {lhs} {operator} {rhs} {{ {then} }} else {{ {otherwise} }}"
-      ),
-      Expression::StringLiteral { string_literal } => write!(f, "{string_literal}"),
-      Expression::Variable { name } => write!(f, "{}", name.lexeme()),
-      Expression::Call { thunk } => write!(f, "{thunk}"),
-      Expression::Group { contents } => write!(f, "({contents})"),
+      } => write!(f, "if {condition} {{ {then} }} else {{ {otherwise} }}"),
+      Self::StringLiteral { string_literal } => write!(f, "{string_literal}"),
+      Self::Variable { name } => write!(f, "{}", name.lexeme()),
+      Self::Call { thunk } => write!(f, "{thunk}"),
+      Self::Group { contents } => write!(f, "({contents})"),
     }
   }
 }
@@ -81,6 +80,13 @@ impl<'src> Serialize for Expression<'src> {
     S: Serializer,
   {
     match self {
+      Self::Assert { condition, error } => {
+        let mut seq: <S as Serializer>::SerializeSeq = serializer.serialize_seq(None)?;
+        seq.serialize_element("assert")?;
+        seq.serialize_element(condition)?;
+        seq.serialize_element(error)?;
+        seq.end()
+      }
       Self::Backtick { contents, .. } => {
         let mut seq = serializer.serialize_seq(None)?;
         seq.serialize_element("evaluate")?;
@@ -103,17 +109,13 @@ impl<'src> Serialize for Expression<'src> {
         seq.end()
       }
       Self::Conditional {
-        lhs,
-        rhs,
+        condition,
         then,
         otherwise,
-        operator,
       } => {
         let mut seq = serializer.serialize_seq(None)?;
         seq.serialize_element("if")?;
-        seq.serialize_element(&operator.to_string())?;
-        seq.serialize_element(lhs)?;
-        seq.serialize_element(rhs)?;
+        seq.serialize_element(condition)?;
         seq.serialize_element(then)?;
         seq.serialize_element(otherwise)?;
         seq.end()
