@@ -4,33 +4,39 @@ use {
     ToKebabCase, ToLowerCamelCase, ToShoutyKebabCase, ToShoutySnakeCase, ToSnakeCase, ToTitleCase,
     ToUpperCamelCase,
   },
+  rand::{seq::SliceRandom, thread_rng},
   semver::{Version, VersionReq},
+  std::collections::HashSet,
   Function::*,
 };
 
 pub(crate) enum Function {
-  Nullary(fn(&FunctionContext) -> Result<String, String>),
-  Unary(fn(&FunctionContext, &str) -> Result<String, String>),
-  UnaryOpt(fn(&FunctionContext, &str, Option<&str>) -> Result<String, String>),
-  Binary(fn(&FunctionContext, &str, &str) -> Result<String, String>),
-  BinaryPlus(fn(&FunctionContext, &str, &str, &[String]) -> Result<String, String>),
-  Ternary(fn(&FunctionContext, &str, &str, &str) -> Result<String, String>),
+  Nullary(fn(&Evaluator) -> Result<String, String>),
+  Unary(fn(&Evaluator, &str) -> Result<String, String>),
+  UnaryOpt(fn(&Evaluator, &str, Option<&str>) -> Result<String, String>),
+  UnaryPlus(fn(&Evaluator, &str, &[String]) -> Result<String, String>),
+  Binary(fn(&Evaluator, &str, &str) -> Result<String, String>),
+  BinaryPlus(fn(&Evaluator, &str, &str, &[String]) -> Result<String, String>),
+  Ternary(fn(&Evaluator, &str, &str, &str) -> Result<String, String>),
 }
 
 pub(crate) fn get(name: &str) -> Option<Function> {
   let function = match name {
     "absolute_path" => Unary(absolute_path),
+    "append" => Binary(append),
     "arch" => Nullary(arch),
     "blake3" => Unary(blake3),
     "blake3_file" => Unary(blake3_file),
-    "canonicalize" => Unary(canonicalize),
     "cache_directory" => Nullary(|_| dir("cache", dirs::cache_dir)),
+    "canonicalize" => Unary(canonicalize),
     "capitalize" => Unary(capitalize),
+    "choose" => Binary(choose),
     "clean" => Unary(clean),
     "config_directory" => Nullary(|_| dir("config", dirs::config_dir)),
     "config_local_directory" => Nullary(|_| dir("local config", dirs::config_local_dir)),
     "data_directory" => Nullary(|_| dir("data", dirs::data_dir)),
     "data_local_directory" => Nullary(|_| dir("local data", dirs::data_local_dir)),
+    "encode_uri_component" => Unary(encode_uri_component),
     "env" => UnaryOpt(env),
     "env_var" => Unary(env_var),
     "env_var_or_default" => Binary(env_var_or_default),
@@ -55,12 +61,14 @@ pub(crate) fn get(name: &str) -> Option<Function> {
     "os_family" => Nullary(os_family),
     "parent_directory" => Unary(parent_directory),
     "path_exists" => Unary(path_exists),
+    "prepend" => Binary(prepend),
     "quote" => Unary(quote),
     "replace" => Ternary(replace),
     "replace_regex" => Ternary(replace_regex),
     "semver_matches" => Binary(semver_matches),
     "sha256" => Unary(sha256),
     "sha256_file" => Unary(sha256_file),
+    "shell" => UnaryPlus(shell),
     "shoutykebabcase" => Unary(shoutykebabcase),
     "shoutysnakecase" => Unary(shoutysnakecase),
     "snakecase" => Unary(snakecase),
@@ -87,6 +95,7 @@ impl Function {
       Nullary(_) => 0..0,
       Unary(_) => 1..1,
       UnaryOpt(_) => 1..2,
+      UnaryPlus(_) => 1..usize::MAX,
       Binary(_) => 2..2,
       BinaryPlus(_) => 2..usize::MAX,
       Ternary(_) => 3..3,
@@ -94,27 +103,36 @@ impl Function {
   }
 }
 
-fn absolute_path(context: &FunctionContext, path: &str) -> Result<String, String> {
-  let abs_path_unchecked = context.search.working_directory.join(path).lexiclean();
+fn absolute_path(evaluator: &Evaluator, path: &str) -> Result<String, String> {
+  let abs_path_unchecked = evaluator.search.working_directory.join(path).lexiclean();
   match abs_path_unchecked.to_str() {
     Some(absolute_path) => Ok(absolute_path.to_owned()),
     None => Err(format!(
       "Working directory is not valid unicode: {}",
-      context.search.working_directory.display()
+      evaluator.search.working_directory.display()
     )),
   }
 }
 
-fn arch(_context: &FunctionContext) -> Result<String, String> {
+fn append(_evaluator: &Evaluator, suffix: &str, s: &str) -> Result<String, String> {
+  Ok(
+    s.split_whitespace()
+      .map(|s| format!("{s}{suffix}"))
+      .collect::<Vec<String>>()
+      .join(" "),
+  )
+}
+
+fn arch(_evaluator: &Evaluator) -> Result<String, String> {
   Ok(target::arch().to_owned())
 }
 
-fn blake3(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn blake3(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(blake3::hash(s.as_bytes()).to_string())
 }
 
-fn blake3_file(context: &FunctionContext, path: &str) -> Result<String, String> {
-  let path = context.search.working_directory.join(path);
+fn blake3_file(evaluator: &Evaluator, path: &str) -> Result<String, String> {
+  let path = evaluator.search.working_directory.join(path);
   let mut hasher = blake3::Hasher::new();
   hasher
     .update_mmap_rayon(&path)
@@ -122,7 +140,7 @@ fn blake3_file(context: &FunctionContext, path: &str) -> Result<String, String> 
   Ok(hasher.finalize().to_string())
 }
 
-fn canonicalize(_context: &FunctionContext, path: &str) -> Result<String, String> {
+fn canonicalize(_evaluator: &Evaluator, path: &str) -> Result<String, String> {
   let canonical =
     std::fs::canonicalize(path).map_err(|err| format!("I/O error canonicalizing path: {err}"))?;
 
@@ -134,7 +152,7 @@ fn canonicalize(_context: &FunctionContext, path: &str) -> Result<String, String
   })
 }
 
-fn capitalize(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn capitalize(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   let mut capitalized = String::new();
   for (i, c) in s.chars().enumerate() {
     if i == 0 {
@@ -146,7 +164,31 @@ fn capitalize(_context: &FunctionContext, s: &str) -> Result<String, String> {
   Ok(capitalized)
 }
 
-fn clean(_context: &FunctionContext, path: &str) -> Result<String, String> {
+fn choose(_evaluator: &Evaluator, n: &str, alphabet: &str) -> Result<String, String> {
+  if alphabet.is_empty() {
+    return Err("empty alphabet".into());
+  }
+
+  let mut chars = HashSet::<char>::with_capacity(alphabet.len());
+
+  for c in alphabet.chars() {
+    if !chars.insert(c) {
+      return Err(format!("alphabet contains repeated character `{c}`"));
+    }
+  }
+
+  let alphabet = alphabet.chars().collect::<Vec<char>>();
+
+  let n = n
+    .parse::<usize>()
+    .map_err(|err| format!("failed to parse `{n}` as positive integer: {err}"))?;
+
+  let mut rng = thread_rng();
+
+  Ok((0..n).map(|_| alphabet.choose(&mut rng).unwrap()).collect())
+}
+
+fn clean(_evaluator: &Evaluator, path: &str) -> Result<String, String> {
   Ok(Path::new(path).lexiclean().to_str().unwrap().to_owned())
 }
 
@@ -166,10 +208,24 @@ fn dir(name: &'static str, f: fn() -> Option<PathBuf>) -> Result<String, String>
   }
 }
 
-fn env_var(context: &FunctionContext, key: &str) -> Result<String, String> {
+fn encode_uri_component(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
+  static PERCENT_ENCODE: percent_encoding::AsciiSet = percent_encoding::NON_ALPHANUMERIC
+    .remove(b'-')
+    .remove(b'_')
+    .remove(b'.')
+    .remove(b'!')
+    .remove(b'~')
+    .remove(b'*')
+    .remove(b'\'')
+    .remove(b'(')
+    .remove(b')');
+  Ok(percent_encoding::utf8_percent_encode(s, &PERCENT_ENCODE).to_string())
+}
+
+fn env_var(evaluator: &Evaluator, key: &str) -> Result<String, String> {
   use std::env::VarError::*;
 
-  if let Some(value) = context.dotenv.get(key) {
+  if let Some(value) = evaluator.dotenv.get(key) {
     return Ok(value.clone());
   }
 
@@ -182,14 +238,10 @@ fn env_var(context: &FunctionContext, key: &str) -> Result<String, String> {
   }
 }
 
-fn env_var_or_default(
-  context: &FunctionContext,
-  key: &str,
-  default: &str,
-) -> Result<String, String> {
+fn env_var_or_default(evaluator: &Evaluator, key: &str, default: &str) -> Result<String, String> {
   use std::env::VarError::*;
 
-  if let Some(value) = context.dotenv.get(key) {
+  if let Some(value) = evaluator.dotenv.get(key) {
     return Ok(value.clone());
   }
 
@@ -202,65 +254,70 @@ fn env_var_or_default(
   }
 }
 
-fn env(context: &FunctionContext, key: &str, default: Option<&str>) -> Result<String, String> {
+fn env(evaluator: &Evaluator, key: &str, default: Option<&str>) -> Result<String, String> {
   match default {
-    Some(val) => env_var_or_default(context, key, val),
-    None => env_var(context, key),
+    Some(val) => env_var_or_default(evaluator, key, val),
+    None => env_var(evaluator, key),
   }
 }
 
-fn error(_context: &FunctionContext, message: &str) -> Result<String, String> {
+fn error(_evaluator: &Evaluator, message: &str) -> Result<String, String> {
   Err(message.to_owned())
 }
 
-fn extension(_context: &FunctionContext, path: &str) -> Result<String, String> {
+fn extension(_evaluator: &Evaluator, path: &str) -> Result<String, String> {
   Utf8Path::new(path)
     .extension()
     .map(str::to_owned)
     .ok_or_else(|| format!("Could not extract extension from `{path}`"))
 }
 
-fn file_name(_context: &FunctionContext, path: &str) -> Result<String, String> {
+fn file_name(_evaluator: &Evaluator, path: &str) -> Result<String, String> {
   Utf8Path::new(path)
     .file_name()
     .map(str::to_owned)
     .ok_or_else(|| format!("Could not extract file name from `{path}`"))
 }
 
-fn file_stem(_context: &FunctionContext, path: &str) -> Result<String, String> {
+fn file_stem(_evaluator: &Evaluator, path: &str) -> Result<String, String> {
   Utf8Path::new(path)
     .file_stem()
     .map(str::to_owned)
     .ok_or_else(|| format!("Could not extract file stem from `{path}`"))
 }
 
-fn invocation_directory(context: &FunctionContext) -> Result<String, String> {
+fn invocation_directory(evaluator: &Evaluator) -> Result<String, String> {
   Platform::convert_native_path(
-    &context.search.working_directory,
-    context.invocation_directory,
+    &evaluator.search.working_directory,
+    &evaluator.config.invocation_directory,
   )
   .map_err(|e| format!("Error getting shell path: {e}"))
 }
 
-fn invocation_directory_native(context: &FunctionContext) -> Result<String, String> {
-  context
+fn invocation_directory_native(evaluator: &Evaluator) -> Result<String, String> {
+  evaluator
+    .config
     .invocation_directory
     .to_str()
     .map(str::to_owned)
     .ok_or_else(|| {
       format!(
         "Invocation directory is not valid unicode: {}",
-        context.invocation_directory.display()
+        evaluator.config.invocation_directory.display()
       )
     })
 }
 
-fn join(
-  _context: &FunctionContext,
-  base: &str,
-  with: &str,
-  and: &[String],
-) -> Result<String, String> {
+fn prepend(_evaluator: &Evaluator, prefix: &str, s: &str) -> Result<String, String> {
+  Ok(
+    s.split_whitespace()
+      .map(|s| format!("{prefix}{s}"))
+      .collect::<Vec<String>>()
+      .join(" "),
+  )
+}
+
+fn join(_evaluator: &Evaluator, base: &str, with: &str, and: &[String]) -> Result<String, String> {
   let mut result = Utf8Path::new(base).join(with);
   for arg in and {
     result.push(arg);
@@ -268,7 +325,7 @@ fn join(
   Ok(result.to_string())
 }
 
-fn just_executable(_context: &FunctionContext) -> Result<String, String> {
+fn just_executable(_evaluator: &Evaluator) -> Result<String, String> {
   let exe_path =
     env::current_exe().map_err(|e| format!("Error getting current executable: {e}"))?;
 
@@ -280,12 +337,12 @@ fn just_executable(_context: &FunctionContext) -> Result<String, String> {
   })
 }
 
-fn just_pid(_context: &FunctionContext) -> Result<String, String> {
+fn just_pid(_evaluator: &Evaluator) -> Result<String, String> {
   Ok(std::process::id().to_string())
 }
 
-fn justfile(context: &FunctionContext) -> Result<String, String> {
-  context
+fn justfile(evaluator: &Evaluator) -> Result<String, String> {
+  evaluator
     .search
     .justfile
     .to_str()
@@ -293,16 +350,16 @@ fn justfile(context: &FunctionContext) -> Result<String, String> {
     .ok_or_else(|| {
       format!(
         "Justfile path is not valid unicode: {}",
-        context.search.justfile.display()
+        evaluator.search.justfile.display()
       )
     })
 }
 
-fn justfile_directory(context: &FunctionContext) -> Result<String, String> {
-  let justfile_directory = context.search.justfile.parent().ok_or_else(|| {
+fn justfile_directory(evaluator: &Evaluator) -> Result<String, String> {
+  let justfile_directory = evaluator.search.justfile.parent().ok_or_else(|| {
     format!(
       "Could not resolve justfile directory. Justfile `{}` had no parent.",
-      context.search.justfile.display()
+      evaluator.search.justfile.display()
     )
   })?;
 
@@ -317,41 +374,41 @@ fn justfile_directory(context: &FunctionContext) -> Result<String, String> {
     })
 }
 
-fn kebabcase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn kebabcase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_kebab_case())
 }
 
-fn lowercamelcase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn lowercamelcase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_lower_camel_case())
 }
 
-fn lowercase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn lowercase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_lowercase())
 }
 
-fn num_cpus(_context: &FunctionContext) -> Result<String, String> {
+fn num_cpus(_evaluator: &Evaluator) -> Result<String, String> {
   let num = num_cpus::get();
   Ok(num.to_string())
 }
 
-fn os(_context: &FunctionContext) -> Result<String, String> {
+fn os(_evaluator: &Evaluator) -> Result<String, String> {
   Ok(target::os().to_owned())
 }
 
-fn os_family(_context: &FunctionContext) -> Result<String, String> {
+fn os_family(_evaluator: &Evaluator) -> Result<String, String> {
   Ok(target::family().to_owned())
 }
 
-fn parent_directory(_context: &FunctionContext, path: &str) -> Result<String, String> {
+fn parent_directory(_evaluator: &Evaluator, path: &str) -> Result<String, String> {
   Utf8Path::new(path)
     .parent()
     .map(Utf8Path::to_string)
     .ok_or_else(|| format!("Could not extract parent directory from `{path}`"))
 }
 
-fn path_exists(context: &FunctionContext, path: &str) -> Result<String, String> {
+fn path_exists(evaluator: &Evaluator, path: &str) -> Result<String, String> {
   Ok(
-    context
+    evaluator
       .search
       .working_directory
       .join(path)
@@ -360,16 +417,16 @@ fn path_exists(context: &FunctionContext, path: &str) -> Result<String, String> 
   )
 }
 
-fn quote(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn quote(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(format!("'{}'", s.replace('\'', "'\\''")))
 }
 
-fn replace(_context: &FunctionContext, s: &str, from: &str, to: &str) -> Result<String, String> {
+fn replace(_evaluator: &Evaluator, s: &str, from: &str, to: &str) -> Result<String, String> {
   Ok(s.replace(from, to))
 }
 
 fn replace_regex(
-  _context: &FunctionContext,
+  _evaluator: &Evaluator,
   s: &str,
   regex: &str,
   replacement: &str,
@@ -382,7 +439,7 @@ fn replace_regex(
   )
 }
 
-fn sha256(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn sha256(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   use sha2::{Digest, Sha256};
   let mut hasher = Sha256::new();
   hasher.update(s);
@@ -390,9 +447,9 @@ fn sha256(_context: &FunctionContext, s: &str) -> Result<String, String> {
   Ok(format!("{hash:x}"))
 }
 
-fn sha256_file(context: &FunctionContext, path: &str) -> Result<String, String> {
+fn sha256_file(evaluator: &Evaluator, path: &str) -> Result<String, String> {
   use sha2::{Digest, Sha256};
-  let path = context.search.working_directory.join(path);
+  let path = evaluator.search.working_directory.join(path);
   let mut hasher = Sha256::new();
   let mut file =
     fs::File::open(&path).map_err(|err| format!("Failed to open `{}`: {err}", path.display()))?;
@@ -402,63 +459,73 @@ fn sha256_file(context: &FunctionContext, path: &str) -> Result<String, String> 
   Ok(format!("{hash:x}"))
 }
 
-fn shoutykebabcase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn shell(evaluator: &Evaluator, command: &str, args: &[String]) -> Result<String, String> {
+  let args = iter::once(command)
+    .chain(args.iter().map(String::as_str))
+    .collect::<Vec<&str>>();
+
+  evaluator
+    .run_command(command, &args)
+    .map_err(|output_error| output_error.to_string())
+}
+
+fn shoutykebabcase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_shouty_kebab_case())
 }
 
-fn shoutysnakecase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn shoutysnakecase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_shouty_snake_case())
 }
 
-fn snakecase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn snakecase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_snake_case())
 }
 
-fn titlecase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn titlecase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_title_case())
 }
 
-fn trim(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn trim(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.trim().to_owned())
 }
 
-fn trim_end(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn trim_end(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.trim_end().to_owned())
 }
 
-fn trim_end_match(_context: &FunctionContext, s: &str, pat: &str) -> Result<String, String> {
+fn trim_end_match(_evaluator: &Evaluator, s: &str, pat: &str) -> Result<String, String> {
   Ok(s.strip_suffix(pat).unwrap_or(s).to_owned())
 }
 
-fn trim_end_matches(_context: &FunctionContext, s: &str, pat: &str) -> Result<String, String> {
+fn trim_end_matches(_evaluator: &Evaluator, s: &str, pat: &str) -> Result<String, String> {
   Ok(s.trim_end_matches(pat).to_owned())
 }
 
-fn trim_start(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn trim_start(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.trim_start().to_owned())
 }
 
-fn trim_start_match(_context: &FunctionContext, s: &str, pat: &str) -> Result<String, String> {
+fn trim_start_match(_evaluator: &Evaluator, s: &str, pat: &str) -> Result<String, String> {
   Ok(s.strip_prefix(pat).unwrap_or(s).to_owned())
 }
 
-fn trim_start_matches(_context: &FunctionContext, s: &str, pat: &str) -> Result<String, String> {
+fn trim_start_matches(_evaluator: &Evaluator, s: &str, pat: &str) -> Result<String, String> {
   Ok(s.trim_start_matches(pat).to_owned())
 }
 
-fn uppercamelcase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn uppercamelcase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_upper_camel_case())
 }
 
-fn uppercase(_context: &FunctionContext, s: &str) -> Result<String, String> {
+fn uppercase(_evaluator: &Evaluator, s: &str) -> Result<String, String> {
   Ok(s.to_uppercase())
 }
 
-fn uuid(_context: &FunctionContext) -> Result<String, String> {
+fn uuid(_evaluator: &Evaluator) -> Result<String, String> {
   Ok(uuid::Uuid::new_v4().to_string())
 }
 
-fn without_extension(_context: &FunctionContext, path: &str) -> Result<String, String> {
+fn without_extension(_evaluator: &Evaluator, path: &str) -> Result<String, String> {
   let parent = Utf8Path::new(path)
     .parent()
     .ok_or_else(|| format!("Could not extract parent from `{path}`"))?;
@@ -473,7 +540,7 @@ fn without_extension(_context: &FunctionContext, path: &str) -> Result<String, S
 /// Check whether a string processes properly as semver (e.x. "0.1.0")
 /// and matches a given semver requirement (e.x. ">=0.1.0")
 fn semver_matches(
-  _context: &FunctionContext,
+  _evaluator: &Evaluator,
   version: &str,
   requirement: &str,
 ) -> Result<String, String> {
@@ -504,7 +571,10 @@ mod tests {
   fn dir_not_unicode() {
     use std::os::unix::ffi::OsStrExt;
     assert_eq!(
-      dir("foo", || Some(OsStr::from_bytes(b"\xe0\x80\x80").into())).unwrap_err(),
+      dir("foo", || Some(
+        std::ffi::OsStr::from_bytes(b"\xe0\x80\x80").into()
+      ))
+      .unwrap_err(),
       "unable to convert foo directory path to string: ���",
     );
   }
