@@ -25,10 +25,10 @@ pub(crate) enum Item<'src> {
 impl<'src> Display for Item<'src> {
   fn fmt(&self, f: &mut Formatter) -> fmt::Result {
     match self {
-      Item::Alias(alias) => write!(f, "{alias}"),
-      Item::Assignment(assignment) => write!(f, "{assignment}"),
-      Item::Comment(comment) => write!(f, "{comment}"),
-      Item::Import {
+      Self::Alias(alias) => write!(f, "{alias}"),
+      Self::Assignment(assignment) => write!(f, "{assignment}"),
+      Self::Comment(comment) => write!(f, "{comment}"),
+      Self::Import {
         relative, optional, ..
       } => {
         write!(f, "import")?;
@@ -39,7 +39,7 @@ impl<'src> Display for Item<'src> {
 
         write!(f, " {relative}")
       }
-      Item::Module {
+      Self::Module {
         name,
         relative,
         optional,
@@ -59,8 +59,8 @@ impl<'src> Display for Item<'src> {
 
         Ok(())
       }
-      Item::Recipe(recipe) => write!(f, "{}", recipe.color_display(Color::never())),
-      Item::Set(set) => write!(f, "{set}"),
+      Self::Recipe(recipe) => write!(f, "{}", recipe.color_display(Color::never())),
+      Self::Set(set) => write!(f, "{set}"),
     }
   }
 }
