@@ -18,10 +18,10 @@ impl<'src> Node<'src> for Ast<'src> {
 impl<'src> Node<'src> for Item<'src> {
   fn tree(&self) -> Tree<'src> {
     match self {
-      Item::Alias(alias) => alias.tree(),
-      Item::Assignment(assignment) => assignment.tree(),
-      Item::Comment(comment) => comment.tree(),
-      Item::Import {
+      Self::Alias(alias) => alias.tree(),
+      Self::Assignment(assignment) => assignment.tree(),
+      Self::Comment(comment) => comment.tree(),
+      Self::Import {
         relative, optional, ..
       } => {
         let mut tree = Tree::atom("import");
@@ -32,7 +32,7 @@ impl<'src> Node<'src> for Item<'src> {
 
         tree.push(format!("{relative}"))
       }
-      Item::Module {
+      Self::Module {
         name,
         optional,
         relative,
@@ -52,8 +52,8 @@ impl<'src> Node<'src> for Item<'src> {
 
         tree
       }
-      Item::Recipe(recipe) => recipe.tree(),
-      Item::Set(set) => set.tree(),
+      Self::Recipe(recipe) => recipe.tree(),
+      Self::Set(set) => set.tree(),
     }
   }
 }
@@ -83,13 +83,19 @@ impl<'src> Node<'src> for Assignment<'src> {
 impl<'src> Node<'src> for Expression<'src> {
   fn tree(&self) -> Tree<'src> {
     match self {
-      Expression::Concatenation { lhs, rhs } => Tree::atom("+").push(lhs.tree()).push(rhs.tree()),
-      Expression::Conditional {
-        lhs,
-        rhs,
+      Self::Assert {
+        condition: Condition { lhs, rhs, operator },
+        error,
+      } => Tree::atom(Keyword::Assert.lexeme())
+        .push(lhs.tree())
+        .push(operator.to_string())
+        .push(rhs.tree())
+        .push(error.tree()),
+      Self::Concatenation { lhs, rhs } => Tree::atom("+").push(lhs.tree()).push(rhs.tree()),
+      Self::Conditional {
+        condition: Condition { lhs, rhs, operator },
         then,
         otherwise,
-        operator,
       } => {
         let mut tree = Tree::atom(Keyword::If.lexeme());
         tree.push_mut(lhs.tree());
@@ -99,7 +105,7 @@ impl<'src> Node<'src> for Expression<'src> {
         tree.push_mut(otherwise.tree());
         tree
       }
-      Expression::Call { thunk } => {
+      Self::Call { thunk } => {
         use Thunk::*;
 
         let mut tree = Tree::atom("call");
@@ -117,6 +123,17 @@ impl<'src> Node<'src> for Expression<'src> {
             tree.push_mut(a.tree());
             if let Some(b) = b.as_ref() {
               tree.push_mut(b.tree());
+            }
+          }
+          UnaryPlus {
+            name,
+            args: (a, rest),
+            ..
+          } => {
+            tree.push_mut(name.lexeme());
+            tree.push_mut(a.tree());
+            for arg in rest {
+              tree.push_mut(arg.tree());
             }
           }
           Binary {
@@ -152,14 +169,14 @@ impl<'src> Node<'src> for Expression<'src> {
 
         tree
       }
-      Expression::Variable { name } => Tree::atom(name.lexeme()),
-      Expression::StringLiteral {
+      Self::Variable { name } => Tree::atom(name.lexeme()),
+      Self::StringLiteral {
         string_literal: StringLiteral { cooked, .. },
       } => Tree::string(cooked),
-      Expression::Backtick { contents, .. } => Tree::atom("backtick").push(Tree::string(contents)),
-      Expression::Group { contents } => Tree::List(vec![contents.tree()]),
-      Expression::Join { lhs: None, rhs } => Tree::atom("/").push(rhs.tree()),
-      Expression::Join {
+      Self::Backtick { contents, .. } => Tree::atom("backtick").push(Tree::string(contents)),
+      Self::Group { contents } => Tree::List(vec![contents.tree()]),
+      Self::Join { lhs: None, rhs } => Tree::atom("/").push(rhs.tree()),
+      Self::Join {
         lhs: Some(lhs),
         rhs,
       } => Tree::atom("/").push(lhs.tree()).push(rhs.tree()),
@@ -252,8 +269,8 @@ impl<'src> Node<'src> for Line<'src> {
 impl<'src> Node<'src> for Fragment<'src> {
   fn tree(&self) -> Tree<'src> {
     match self {
-      Fragment::Text { token } => Tree::string(token.lexeme()),
-      Fragment::Interpolation { expression } => Tree::List(vec![expression.tree()]),
+      Self::Text { token } => Tree::string(token.lexeme()),
+      Self::Interpolation { expression } => Tree::List(vec![expression.tree()]),
     }
   }
 }
@@ -265,6 +282,7 @@ impl<'src> Node<'src> for Set<'src> {
 
     match &self.value {
       Setting::AllowDuplicateRecipes(value)
+      | Setting::AllowDuplicateVariables(value)
       | Setting::DotenvLoad(value)
       | Setting::Export(value)
       | Setting::Fallback(value)
