@@ -2,27 +2,48 @@ use super::*;
 
 #[derive(Debug)]
 pub(crate) struct Scope<'src: 'run, 'run> {
-  parent: Option<&'run Scope<'src, 'run>>,
+  parent: Option<&'run Self>,
   bindings: Table<'src, Binding<'src, String>>,
 }
 
 impl<'src, 'run> Scope<'src, 'run> {
-  pub(crate) fn child(&'run self) -> Scope<'src, 'run> {
+  pub(crate) fn child(&'run self) -> Self {
     Self {
       parent: Some(self),
       bindings: Table::new(),
     }
   }
 
-  pub(crate) fn new() -> Scope<'src, 'run> {
-    Self {
+  pub(crate) fn root() -> Self {
+    let mut root = Self {
       parent: None,
       bindings: Table::new(),
+    };
+
+    for (key, value) in constants() {
+      root.bind(
+        false,
+        Name {
+          token: Token {
+            column: 0,
+            kind: TokenKind::Identifier,
+            length: key.len(),
+            line: 0,
+            offset: 0,
+            path: Path::new("PRELUDE"),
+            src: key,
+          },
+        },
+        (*value).into(),
+      );
     }
+
+    root
   }
 
   pub(crate) fn bind(&mut self, export: bool, name: Name<'src>, value: String) {
     self.bindings.insert(Binding {
+      depth: 0,
       export,
       name,
       value,
@@ -49,7 +70,7 @@ impl<'src, 'run> Scope<'src, 'run> {
     self.bindings.keys().copied()
   }
 
-  pub(crate) fn parent(&self) -> Option<&'run Scope<'src, 'run>> {
+  pub(crate) fn parent(&self) -> Option<&'run Self> {
     self.parent
   }
 }
