@@ -1,4 +1,9 @@
-use super::*;
+use {
+  super::*,
+  clap_mangen::Man,
+  std::io::{Read, Seek},
+  tempfile::tempfile,
+};
 
 const INIT_JUSTFILE: &str = "default:\n    echo 'Hello, world!'\n";
 
@@ -15,7 +20,7 @@ pub(crate) enum Subcommand {
     overrides: BTreeMap<String, String>,
   },
   Completions {
-    shell: String,
+    shell: clap_complete::Shell,
   },
   Dump,
   Edit,
@@ -26,6 +31,7 @@ pub(crate) enum Subcommand {
   Format,
   Init,
   List,
+  Man,
   Run {
     arguments: Vec<String>,
     overrides: BTreeMap<String, String>,
@@ -50,8 +56,9 @@ impl Subcommand {
         Self::changelog();
         return Ok(());
       }
-      Completions { shell } => return Self::completions(shell),
+      Completions { shell } => return Self::completions(*shell),
       Init => return Self::init(config),
+      Man => return Self::man(),
       Run {
         arguments,
         overrides,
@@ -83,7 +90,7 @@ impl Subcommand {
       Show { ref name } => Self::show(config, name, justfile)?,
       Summary => Self::summary(config, justfile),
       Variables => Self::variables(justfile),
-      Changelog | Completions { .. } | Edit | Init | Run { .. } => unreachable!(),
+      Changelog | Completions { .. } | Edit | Init | Man | Run { .. } => unreachable!(),
     }
 
     Ok(())
@@ -213,10 +220,7 @@ impl Subcommand {
       return Err(Error::NoChoosableRecipes);
     }
 
-    let chooser = chooser
-      .map(OsString::from)
-      .or_else(|| env::var_os(config::CHOOSER_ENVIRONMENT_KEY))
-      .unwrap_or_else(|| config::chooser_default(&search.justfile));
+    let chooser = chooser.map_or_else(|| config::chooser_default(&search.justfile), From::from);
 
     let result = justfile
       .settings
@@ -275,8 +279,8 @@ impl Subcommand {
     justfile.run(config, search, overrides, &recipes)
   }
 
-  fn completions(shell: &str) -> RunResult<'static, ()> {
-    use clap::Shell;
+  fn completions(shell: clap_complete::Shell) -> RunResult<'static, ()> {
+    use clap_complete::Shell;
 
     fn replace(haystack: &mut String, needle: &str, replacement: &str) -> RunResult<'static, ()> {
       if let Some(index) = haystack.find(needle) {
@@ -289,15 +293,28 @@ impl Subcommand {
       }
     }
 
-    let shell = shell
-      .parse::<Shell>()
-      .expect("Invalid value for clap::Shell");
+    let mut script = {
+      let mut tempfile = tempfile().map_err(|io_error| Error::TempfileIo { io_error })?;
 
-    let buffer = Vec::new();
-    let mut cursor = Cursor::new(buffer);
-    Config::app().gen_completions_to(env!("CARGO_PKG_NAME"), shell, &mut cursor);
-    let buffer = cursor.into_inner();
-    let mut script = String::from_utf8(buffer).expect("Clap completion not UTF-8");
+      clap_complete::generate(
+        shell,
+        &mut crate::config::Config::app(),
+        env!("CARGO_PKG_NAME"),
+        &mut tempfile,
+      );
+
+      tempfile
+        .rewind()
+        .map_err(|io_error| Error::TempfileIo { io_error })?;
+
+      let mut buffer = String::new();
+
+      tempfile
+        .read_to_string(&mut buffer)
+        .map_err(|io_error| Error::TempfileIo { io_error })?;
+
+      buffer
+    };
 
     match shell {
       Shell::Bash => {
@@ -319,7 +336,7 @@ impl Subcommand {
           replace(&mut script, needle, replacement)?;
         }
       }
-      Shell::Elvish => {}
+      _ => {}
     }
 
     println!("{}", script.trim());
@@ -426,23 +443,51 @@ impl Subcommand {
     }
   }
 
-  fn list(config: &Config, level: usize, justfile: &Justfile) {
-    // Construct a target to alias map.
-    let mut recipe_aliases: BTreeMap<&str, Vec<&str>> = BTreeMap::new();
-    for alias in justfile.aliases.values() {
-      if alias.is_private() {
-        continue;
-      }
+  fn man() -> Result<(), Error<'static>> {
+    let mut buffer = Vec::<u8>::new();
 
-      if recipe_aliases.contains_key(alias.target.name.lexeme()) {
-        let aliases = recipe_aliases.get_mut(alias.target.name.lexeme()).unwrap();
-        aliases.push(alias.name.lexeme());
-      } else {
-        recipe_aliases.insert(alias.target.name.lexeme(), vec![alias.name.lexeme()]);
+    Man::new(Config::app())
+      .render(&mut buffer)
+      .expect("writing to buffer cannot fail");
+
+    let mut stdout = io::stdout().lock();
+
+    stdout
+      .write_all(&buffer)
+      .map_err(|io_error| Error::StdoutIo { io_error })?;
+
+    stdout
+      .flush()
+      .map_err(|io_error| Error::StdoutIo { io_error })?;
+
+    Ok(())
+  }
+
+  fn list(config: &Config, level: usize, justfile: &Justfile) {
+    const MAX_WIDTH: usize = 50;
+
+    if level == 0 {
+      print!("{}", config.list_heading);
+    }
+
+    // Construct a target to alias map.
+    let mut recipe_aliases = BTreeMap::<&str, Vec<&str>>::new();
+    if !config.no_aliases {
+      for alias in justfile.aliases.values() {
+        if alias.is_private() {
+          continue;
+        }
+
+        if recipe_aliases.contains_key(alias.target.name.lexeme()) {
+          let aliases = recipe_aliases.get_mut(alias.target.name.lexeme()).unwrap();
+          aliases.push(alias.name.lexeme());
+        } else {
+          recipe_aliases.insert(alias.target.name.lexeme(), vec![alias.name.lexeme()]);
+        }
       }
     }
 
-    let mut line_widths: BTreeMap<&str, usize> = BTreeMap::new();
+    let mut line_widths = BTreeMap::<&str, usize>::new();
 
     for (name, recipe) in &justfile.recipes {
       if !recipe.is_public() {
@@ -450,26 +495,24 @@ impl Subcommand {
       }
 
       for name in iter::once(name).chain(recipe_aliases.get(name).unwrap_or(&Vec::new())) {
-        let mut line_width = UnicodeWidthStr::width(*name);
-
-        for parameter in &recipe.parameters {
-          line_width += UnicodeWidthStr::width(
-            format!(" {}", parameter.color_display(Color::never())).as_str(),
-          );
-        }
-
-        if line_width <= 30 {
-          line_widths.insert(name, line_width);
-        }
+        line_widths.insert(
+          name,
+          UnicodeWidthStr::width(
+            RecipeSignature { name, recipe }
+              .color_display(Color::never())
+              .to_string()
+              .as_str(),
+          ),
+        );
       }
     }
 
-    let max_line_width = cmp::min(line_widths.values().copied().max().unwrap_or(0), 30);
-    let doc_color = config.color.stdout().doc();
-
-    if level == 0 {
-      print!("{}", config.list_heading);
-    }
+    let max_line_width = line_widths
+      .values()
+      .filter(|line_width| **line_width <= MAX_WIDTH)
+      .copied()
+      .max()
+      .unwrap_or_default();
 
     for recipe in justfile.public_recipes(config.unsorted) {
       let name = recipe.name();
@@ -478,39 +521,34 @@ impl Subcommand {
         .chain(recipe_aliases.get(name).unwrap_or(&Vec::new()))
         .enumerate()
       {
-        print!("{}{name}", config.list_prefix.repeat(level + 1));
-        for parameter in &recipe.parameters {
-          print!(" {}", parameter.color_display(config.color.stdout()));
-        }
+        print!(
+          "{}{}",
+          config.list_prefix.repeat(level + 1),
+          RecipeSignature { name, recipe }.color_display(config.color.stdout())
+        );
 
-        // Declaring this outside of the nested loops will probably be more efficient,
-        // but it creates all sorts of lifetime issues with variables inside the loops.
-        // If this is inlined like the docs say, it shouldn't make any difference.
-        let print_doc = |doc| {
+        let doc = match (i, recipe.doc) {
+          (0, Some(doc)) => Some(Cow::Borrowed(doc)),
+          (0, None) => None,
+          _ => Some(Cow::Owned(format!("alias for `{}`", recipe.name))),
+        };
+
+        if let Some(doc) = doc {
           print!(
             " {:padding$}{} {}",
             "",
-            doc_color.paint("#"),
-            doc_color.paint(doc),
-            padding = max_line_width
-              .saturating_sub(line_widths.get(name).copied().unwrap_or(max_line_width))
+            config.color.stdout().doc().paint("#"),
+            config.color.stdout().doc().paint(&doc),
+            padding = max_line_width.saturating_sub(line_widths[name]),
           );
-        };
-
-        match (i, recipe.doc) {
-          (0, Some(doc)) => print_doc(doc),
-          (0, None) => (),
-          _ => {
-            let alias_doc = format!("alias for `{}`", recipe.name);
-            print_doc(&alias_doc);
-          }
         }
+
         println!();
       }
     }
 
     for (name, module) in &justfile.modules {
-      println!("    {name}:");
+      println!("{}{name}:", config.list_prefix.repeat(level + 1));
       Self::list(config, level + 1, module);
     }
   }
