@@ -10,7 +10,7 @@ pub(crate) fn config(args: &[&str]) -> Config {
 
   let app = Config::app();
 
-  let matches = app.get_matches_from_safe(args).unwrap();
+  let matches = app.try_get_matches_from(args).unwrap();
 
   Config::from_matches(&matches).unwrap()
 }
@@ -60,6 +60,7 @@ pub(crate) fn analysis_error(
   let tokens = Lexer::test_lex(src).expect("Lexing failed in parse test...");
 
   let ast = Parser::parse(
+    0,
     &PathBuf::new(),
     &Namepath::default(),
     0,
@@ -88,7 +89,7 @@ pub(crate) fn analysis_error(
           length,
           path: "justfile".as_ref(),
         },
-        kind: Box::new(kind),
+        kind: kind.into(),
       };
       assert_eq!(have, want);
     }
