@@ -200,7 +200,11 @@ fn modules_use_module_settings() {
   Test::new()
     .write(
       "foo.just",
-      "set allow-duplicate-recipes\nfoo:\nfoo:\n @echo FOO\n",
+      "set allow-duplicate-recipes
+foo:
+foo:
+  @echo FOO
+",
     )
     .justfile(
       "
@@ -215,7 +219,13 @@ fn modules_use_module_settings() {
     .run();
 
   Test::new()
-    .write("foo.just", "\nfoo:\nfoo:\n @echo FOO\n")
+    .write(
+      "foo.just",
+      "foo:
+foo:
+  @echo FOO
+",
+    )
     .justfile(
       "
         mod foo
@@ -230,10 +240,10 @@ fn modules_use_module_settings() {
     .arg("foo")
     .stderr(
       "
-      error: Recipe `foo` first defined on line 2 is redefined on line 3
-       ——▶ foo.just:3:1
+      error: Recipe `foo` first defined on line 1 is redefined on line 2
+       ——▶ foo.just:2:1
         │
-      3 │ foo:
+      2 │ foo:
         │ ^^^
     ",
     )
@@ -503,28 +513,6 @@ fn missing_optional_modules_do_not_conflict() {
 }
 
 #[test]
-fn list_displays_recipes_in_submodules() {
-  Test::new()
-    .write("foo.just", "bar:\n @echo FOO")
-    .justfile(
-      "
-        mod foo
-      ",
-    )
-    .test_round_trip(false)
-    .arg("--unstable")
-    .arg("--list")
-    .stdout(
-      "
-      Available recipes:
-          foo:
-              bar
-    ",
-    )
-    .run();
-}
-
-#[test]
 fn root_dotenv_is_available_to_submodules() {
   Test::new()
     .write("foo.just", "foo:\n @echo $DOTENV_KEY")
@@ -682,37 +670,6 @@ fn module_paths_beginning_with_tilde_are_expanded_to_homdir() {
     .arg("foo")
     .stdout("FOOBAR\n")
     .env("HOME", "foobar")
-    .run();
-}
-
-#[test]
-fn module_recipe_list_alignment_ignores_private_recipes() {
-  Test::new()
-    .write(
-      "foo.just",
-      "
-# foos
-foo:
- @echo FOO
-
-[private]
-barbarbar:
- @echo BAR
-
-@_bazbazbaz:
- @echo BAZ
-      ",
-    )
-    .justfile("mod foo")
-    .test_round_trip(false)
-    .arg("--unstable")
-    .arg("--list")
-    .stdout(
-      "Available recipes:
-    foo:
-        foo # foos
-",
-    )
     .run();
 }
 
